[Fix] 메뉴판 API 버그 수정

+++ src/main/java/com/ourmenu/backend/domain/menu/application/MenuFolderService.java
@@ -12,6 +12,7 @@ import com.ourmenu.backend.domain.menu.exception.NotFoundMenuFolderException;
 import com.ourmenu.backend.domain.menu.exception.OutOfBoundCustomIndexException;
 import com.ourmenu.backend.domain.menu.util.DefaultImgConverter;
 import java.util.List;
+import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -33,10 +34,10 @@ public class MenuFolderService {
      */
     @Transactional
     public SaveMenuFolderResponse saveMenuFolder(MenuFolderDto menuFolderDto) {
-        String menuFolderImgUrl = null;
-        if (menuFolderDto.getMenuFolderImg() != null) {
-            menuFolderImgUrl = awsS3Service.uploadFileAsync(menuFolderDto.getMenuFolderImg());
-        }
+        String menuFolderImgUrl = Optional.ofNullable(menuFolderDto.getMenuFolderImg())
+                .map(awsS3Service::uploadFileAsync)
+                .orElse(null);
+
         MenuFolder menuFolder = saveMenuFolder(menuFolderDto, menuFolderImgUrl);
         return SaveMenuFolderResponse.of(menuFolder, menuFolderDto.getMenuIds(),
                 defaultImgConvertor.getDefaultMenuFolderImgUrl());
@@ -54,29 +55,39 @@ public class MenuFolderService {
         menuFolderRepository.delete(menuFolder);
         int maxIndex = menuFolderRepository.findMaxIndex();
 
-        if (menuFolderId != maxIndex) {
-            menuFolderRepository.decrementIndexes(userId, (int) (menuFolderId + 1), maxIndex);
+        if (menuFolder.getIndex() != maxIndex) {
+            menuFolderRepository.decrementIndexes(userId, (menuFolder.getIndex() + 1), maxIndex);
         }
-        awsS3Service.deleteFileAsync(menuFolder.getImgUrl());
+
+        if (menuFolder.getImgUrl() != null) {
+            awsS3Service.deleteFileAsync(menuFolder.getImgUrl());
+        }
     }
 
     @Transactional
-    public UpdateMenuFolderResponse updateMenuFolder(Long userId, Long menuFolderId, MenuFolderDto menuFolderDto) {
+    public UpdateMenuFolderResponse updateMenuFolder(Long userId, Long menuFolderId, MenuFolderDto menuFolderDto,
+                                                     Boolean isImageModified) {
 
         MenuFolder menuFolder = findOne(userId, menuFolderId);
         if (menuFolderDto.getMenuIds() != null) {
             menuMenuFolderService.updateMenuMenuFolder(userId, menuFolderId, menuFolderDto.getMenuIds());
         }
 
-        if (menuFolderDto.getMenuFolderImg() != null) {
-            String imgUrl = awsS3Service.uploadFileAsync(menuFolderDto.getMenuFolderImg());
-            menuFolder.update(menuFolderDto, imgUrl);
-            List<MenuMenuFolder> menuMenuFolders = menuMenuFolderService.findAllByMenuFolderId(menuFolderId);
-            return UpdateMenuFolderResponse.of(menuFolder, menuMenuFolders,
-                    defaultImgConvertor.getDefaultMenuFolderImgUrl());
-        }
+        if (isImageModified) {
+            //기존에 있는 이미지는 삭제한다
+            if (menuFolder.getImgUrl() != null) {
+                awsS3Service.deleteFileAsync(menuFolder.getImgUrl());
+                menuFolder.updateImg(null);
+            }
 
-        menuFolder.update(menuFolderDto, null);
+            //추가할 이미지가 있으면 추가한다
+            if (menuFolderDto.getMenuFolderImg() != null) {
+                String imgUrl = awsS3Service.uploadFileAsync(menuFolderDto.getMenuFolderImg());
+
+                menuFolder.updateImg(imgUrl);
+            }
+        }
+        menuFolder.update(menuFolderDto);
         List<MenuMenuFolder> menuMenuFolders = menuMenuFolderService.findAllByMenuFolderId(menuFolderId);
         return UpdateMenuFolderResponse.of(menuFolder, menuMenuFolders,
                 defaultImgConvertor.getDefaultMenuFolderImgUrl());
